Update per-slot processing to v0.4.0
